moved sales business logic to models package
